update /accounts api path into /users and /auth
